Syncing master from coninalpha

+++ hummingbot/strategy/pure_market_making/constant_size_sizing_delegate.pyx
@@ -17,7 +17,7 @@ s_decimal_0 = Decimal(0)
 
 cdef class ConstantSizeSizingDelegate(OrderSizingDelegate):
 
-    def __init__(self, order_size: float):
+    def __init__(self, order_size: Decimal):
         super().__init__()
         self._order_size = order_size
         self._log_warning_order_size = True
@@ -31,7 +31,7 @@ cdef class ConstantSizeSizingDelegate(OrderSizingDelegate):
         return s_logger
 
     @property
-    def order_size(self) -> float:
+    def order_size(self) -> Decimal:
         return self._order_size
 
     cdef object c_get_order_size_proposal(self,
@@ -42,10 +42,10 @@ cdef class ConstantSizeSizingDelegate(OrderSizingDelegate):
         cdef:
             MarketBase market = market_info.market
             object buy_fees
-            object base_asset_balance = Decimal(market.c_get_available_balance(market_info.base_asset))
-            object quote_asset_balance = Decimal(market.c_get_available_balance(market_info.quote_asset))
-            object bid_order_size = Decimal(self._order_size)
-            object ask_order_size = Decimal(self._order_size)
+            object base_asset_balance = market.c_get_available_balance(market_info.base_asset)
+            object quote_asset_balance = market.c_get_available_balance(market_info.quote_asset)
+            object bid_order_size = self._order_size
+            object ask_order_size = self._order_size
             object quantized_bid_order_size
             object quantized_ask_order_size
             bint has_active_bid = False
@@ -81,9 +81,7 @@ cdef class ConstantSizeSizingDelegate(OrderSizingDelegate):
                                         quantized_bid_order_size,
                                         pricing_proposal.buy_order_prices[0])
 
-            required_quote_asset_balance = (pricing_proposal.buy_order_prices[0] *
-                                            Decimal(1 + buy_fees.percent) *
-                                            quantized_bid_order_size)
+            required_quote_asset_balance = pricing_proposal.buy_order_prices[0] * (1 + buy_fees.percent) * quantized_bid_order_size
 
         if self._log_warning_order_size:
             if quantized_bid_order_size == s_decimal_0:
